+++ lib/views/dashboard/widgets/change_server_button.dart
@@ -74,7 +74,7 @@ class ChangeServerButton extends ConsumerWidget {
     }
 
     final serverInfoGroupName = _decodeBase64IfNeeded(
-      profile.providerHeaders['flclashx-serverinfo'],
+      profile.providerHeaders['dropweb-serverinfo'],
     );
 
     if (serverInfoGroupName == null || serverInfoGroupName.isEmpty) {